assertions.length: reinserted `_NoLengthError()` into `_get_length()`
I originally abstracted out _NoLengthError, because it was used twice.

Then I realized there was a block of code that was being used twice, so I factored that out, too.

Then I forgot that when I factored that out, the stuff I originally factored out didn’t need to be factored out anymore, so I factored it back in.

That’s a fact.  (or). :P